backend/requirements.txt
fastapi
uvicorn
pandas
groq
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-dotenv
PyJWT
bcrypt==4.0.1
passlib[bcrypt]==1.7.4